Update M1HW_JeremyLocklear.py
Changed the way the calculator works, now it gets the user input and goes and takes the numbers into a different function and calculates. This way for the time being when future updates are introduced the functions can be condensed into a class called Math. From there it will handle the calculations and such.