Refactored and Renamed.
Renamed the old basic calculator file for importing into multiple files. To import into the main file that will handle the main functions for this calculator. Rather than simply having two menus within the same function.

Had to recreate the whole basic calculator core after realizing that the original way after refactoring the code would not be suitable to the rest of the primary functions that would have been needed to continue.

So now the code will run off of seperate files that will all be imported into one main file. This is the start of the new refactored coded.