Updates and Changes 1.2
Basically scrapped the idea of multiple files as it was causing too much of an issue to create. Was forced to look at some base code that was already created. To try and fix the fundamentally broken code from trying to separate the code into different files. Trying to make classes for them was too much of a hassle to commit to. So instead the code has been re refactored to fit in a single file. With different functions all handling their one main thing and trying not to do others. This took time and learning experience that was very much needed and helps knowing that you can rely on others in case something does go wrong.